feat: add types GaVector, GaBivector, GaRotor
These are all wrappers around GaMultivector.
This makes sure that when using rotation function people give the
correct types of multivectors with only the needed element.
These new types mostly have all defined interactions between the diffent types.

The only outlier is GaRotor that only need interactions between it self
and GaVectors.

These types have been tested with access to the std lib and need to be
tested on a microcontroller.